refactor: Int 대신 Money 사용

+++ src/main/kotlin/lotto/ui/ResultView.kt
@@ -21,7 +21,7 @@ object ResultView {
         Rank.values().forEach {
             println("${it.count}개 일치 (${it.winningAmount}원) - ${matchResult.countOf(it)}개")
         }
-        val rateOfReturn = matchResult.rateOfReturn(money.intValue)
+        val rateOfReturn = matchResult.rateOfReturn(money)
         val benefitOutcome = BenefitOutcome.from(rateOfReturn)
         println("총 수익률은 $rateOfReturn 입니다.(기준이 1이기 때문에 결과적으로 ${benefitOutcome.description}라는 의미임)")
     }
